之前IIC读取错误是中文注释的原因
现在已经改好了，明天测试姿态解算

+++ MPU_6050.h
@@ -147,7 +147,7 @@
 
 typedef struct 
 {
-	int16_t Ax,Ay,Az,Tt,Gx,Gy,Gz;
+	float Ax,Ay,Az,Tt,Gx,Gy,Gz;
 }	MPU6050_Struct;
 
 #endif
